+++ src/main/java/org/socialculture/platform/ticket/repository/querydsl/TicketRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package org.socialculture.platform.ticket.repository;
+package org.socialculture.platform.ticket.repository.querydsl;
 
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
@@ -9,6 +9,7 @@ import org.socialculture.platform.global.apiResponse.exception.GeneralException;
 import org.socialculture.platform.performance.entity.QPerformanceEntity;
 import org.socialculture.platform.ticket.entity.QTicketEntity;
 import org.socialculture.platform.ticket.entity.TicketEntity;
+import org.socialculture.platform.ticket.repository.TicketRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -19,40 +20,21 @@ import java.util.Objects;
  *
  * @author ycjung
  */
-@Repository
 @RequiredArgsConstructor
-public class DslTicketRepositoryImpl implements DslTicketRepository {
+public class TicketRepositoryCustomImpl implements TicketRepositoryCustom {
 
     private final JPAQueryFactory jpaQueryFactory;
 
     /**
-     * fetchJoin 을 통한 나의 티켓 전체 조회
-     * @param memberId
+     * fetchJoin 을 통한 나의 티켓 전체 조회 - 페이징 + 소트
+     * @param email
+     * @param offset
+     * @param pageSize
+     * @param sortOption
      * @return
      */
     @Override
-    public List<TicketEntity> getAllTicketsByMemberId(long memberId) {
-        QTicketEntity ticketEntity = QTicketEntity.ticketEntity;
-
-        return jpaQueryFactory.selectFrom(ticketEntity)
-                .join(ticketEntity.performance, QPerformanceEntity.performanceEntity).fetchJoin() // join fetch 사용
-                .where(ticketEntity.member.memberId.eq(memberId))
-                .fetch();
-    }
-
-    @Override
-    public List<TicketEntity> getAllTicketsByMemberIdWithPage(long memberId, long offset, int pageSize) {
-        QTicketEntity ticketEntity = QTicketEntity.ticketEntity;
-
-        return jpaQueryFactory.selectFrom(ticketEntity)
-                .join(ticketEntity.performance, QPerformanceEntity.performanceEntity).fetchJoin()
-                .offset(offset) // 조회하려고 하는 페이지가 과거일수록 느려진다.
-                .limit(pageSize)
-                .fetch();
-    }
-
-    @Override
-    public List<TicketEntity> getAllTicketsByMemberIdWithPageAndSortOptionDesc(long memberId, long offset, int pageSize, String sortOption) {
+    public List<TicketEntity> getAllTicketsByEmailWithPageAndSortOptionDesc(String email, long offset, int pageSize, String sortOption) {
         QTicketEntity ticketEntity = QTicketEntity.ticketEntity;
         OrderSpecifier<?> orderSpecifier;
 
@@ -68,6 +50,7 @@ public class DslTicketRepositoryImpl implements DslTicketRepository {
 
         return jpaQueryFactory.selectFrom(ticketEntity)
                 .join(ticketEntity.performance, QPerformanceEntity.performanceEntity).fetchJoin()
+                .where(ticketEntity.member.email.eq(email))
                 .offset(offset) // 조회하려고 하는 페이지가 과거일수록 느려진다.
                 .limit(pageSize)
                 .orderBy(orderSpecifier)
